Extract yash_semantics::trap::signal module
This commit moves the run_traps_for_caught_signals function function to
the new module, so that we have only submodules under the trap module.
